Added choices and fixed localStorage issues.
The choices that come from the API are separated, hence, inside getData(), each element was mapped to create a new key-value pair that contains all choices that are shuffled for that question.

For the question state, an isLoaded key-value pair was added to show a loading screen when the data from the API is not yet fetched.

For getData, try{} and catch{} were added.

Under displayQuiz, the choices were mapped to shorten the code instead of the former hard coded <input> elements.

initializeStorage is now fixed by adding 'window' before localStorage.setItem().